# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para gestión de crédito y cobranza en Venezuela
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventType } from './activityEventType';

export interface ActivityEvent {
  id: string;
  type: ActivityEventType;
  title: string;
  description: string;
  /** @nullable */
  amount: number | null;
  occurredAt: Date;
}