# lib/api-zod/src/generated/types/creditDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para gestión de crédito y cobranza en Venezuela
 * OpenAPI spec version: 0.1.0
 */
import type { Credit } from './credit';
import type { Installment } from './installment';

export type CreditDetail = Credit & {
  installments: Installment[];
};