# lib/api-zod/src/generated/types/creditInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para gestión de crédito y cobranza en Venezuela
 * OpenAPI spec version: 0.1.0
 */

export interface CreditInput {
  customerId: string;
  /** @minLength 1 */
  productName: string;
  /** @exclusiveMinimum 0 */
  totalAmount: number;
  /** @minimum 0 */
  downPayment: number;
  /**
     * @minimum 2
     * @maximum 20
     */
  installmentCount: number;
  firstDueDate: Date;
}