# lib/api-zod/src/generated/types/customerStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para gestión de crédito y cobranza en Venezuela
 * OpenAPI spec version: 0.1.0
 */

export type CustomerStatus = typeof CustomerStatus[keyof typeof CustomerStatus];


export const CustomerStatus = {
  current: 'current',
  overdue: 'overdue',
  finished: 'finished',
} as const;