# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para gestión de crédito y cobranza en Venezuela
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalToCollect: number;
  collectedThisMonth: number;
  overdueAmount: number;
  activeCustomers: number;
  overdueCustomers: number;
  dueToday: number;
}