# lib/api-zod/src/generated/types/exchangeRate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para gestión de crédito y cobranza en Venezuela
 * OpenAPI spec version: 0.1.0
 */
import type { ExchangeRateSource } from './exchangeRateSource';

export interface ExchangeRate {
  /** @exclusiveMinimum 0 */
  value: number;
  source: ExchangeRateSource;
  updatedAt: Date;
}