# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para gestión de crédito y cobranza en Venezuela
 * OpenAPI spec version: 0.1.0
 */

export type PaymentMethod = typeof PaymentMethod[keyof typeof PaymentMethod];


export const PaymentMethod = {
  mobile_payment: 'mobile_payment',
  cash_usd: 'cash_usd',
  zelle: 'zelle',
  binance: 'binance',
} as const;